+++ src/components/Datatable/Table/TableRow/TableRow.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import styled from 'styled-components';
+
+const StyledTableRow = styled.tr``;
+
+const TableRow: React.FC = ({ children, ...props }) => (
+  <StyledTableRow {...props}>{children}</StyledTableRow>
+);
+
+export default TableRow;